feat: metadata + noindex on public auth/utility pages (#8, scoped)
Scoped #8 to genuinely public pages (per decision to skip auth-gated
pages no crawler ever sees). The high-value public pages - home,
products, product detail, category, about, contact, search, faqs,
privacy, terms - already export metadata via existing server wrappers.
The gap among public pages was the four client-component utility pages
that couldn't export metadata: login, register, forgot-password,
unsubscribe.

Split each into a server page.js wrapper (exports metadata) + the
existing client form component. Added a proper <title> and, since these
are utility pages that shouldn't rank, robots: { index: false }.

Deliberately NOT touched: the ~19 auth-gated client pages (cart, orders,
wishlist, checkout, all dashboard and user-dashboard pages,
edit-product). They're never indexed, so metadata wrappers would be
near-identical dead files - skipped per the public-pages-only scope.

Verified: /login, /register, /forgot-password, /unsubscribe render with
correct titles and noindex, forms still work.
